pyproject.toml
[tool.poetry]
name = "prefect-auth-proxy"
version = "0.1.0"
description = "Authentication and proxy for Prefect Server"
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/prefeitura-rio/prefect-auth-proxy"
repository = "https://github.com/prefeitura-rio/prefect-auth-proxy"
keywords = ["prefect", "proxy", "authentication", "authorization"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
uvicorn = {extras = ["standard"], version = "^0.23.1"}
loguru = "^0.7.0"
aerich = "^0.7.2"
passlib = "^1.7.4"
pendulum = "^2.1.2"
httpx = "^0.24.1"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
pre-commit = "^2.18.1"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
taskipy = "^1.11.0"
pytest-ordering = "^0.6"
isort = "^5.12.0"
semgrep = "^1.34.1"

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line-length = 100
target-version = ["py310"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["dist", "build", ".tox", ".eggs", "scripts"]
addopts = ["--doctest-modules", "--cov=app", "-ra", "-v"]

[tool.taskipy.tasks]
create_user = "python scripts/create_user.py"
lint = "semgrep scan --error --verbose --config auto && black . && isort . && flake8 ."
serve = "uvicorn app.main:app --reload --port 8001"
test = "pytest"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
